backend/atomic-services/swap-request/requirements.txt
flask
flask_sqlalchemy
marshmallow
python-dotenv
sqlalchemy
psycopg2-binary
pika
